requirements/requirements_base.txt
Django>=2.0.2

django-filter>=1.1.0
Markdown>=2.6.11
djangorestframework>=3.7.7
djangorestframework-jwt>=1.11.0

celery>=4.1.0
kombu>=4.1.0
billiard>=3.5.0.3
amqp>=2.2.2
redis>=2.10.6
vine>=1.1.4
pytz>=2018.3

asgiref>=2.1.6
asgi-redis>=1.4.3
channels>=2.0.2
channels-redis>=2.0.3
async-timeout>=2.0.0